Fix regression on assertCommandSuccess.
